[v1.1.35] Options Submenu + New attributes Json
Options Submenu available: user can choose if he erases or not the
database when he deletes the plugin.
"Tags" and "alignment" for the JSON badges.
New php class Badge to simplify the code.

+++ uninstall.php
@@ -22,7 +22,7 @@ delete_option( $option_name );
 delete_site_option( $option_name );  
  
 $delDb = get_option('badge_delete_db');	
-//if( $delDb == 'checked="checked"' and get_option('badge_delete_db_confirm') == 'yes') {
+if($delDb == 1) {
     //Drop all the 'badges4languages-plugin' Database Tables
     global $wpdb;
     $wpdb->query( "DROP TABLE IF EXISTS {$wpdb->prefix}b4l_languages" );
@@ -35,7 +35,7 @@ $delDb = get_option('badge_delete_db');
     $wpdb->query( "DROP TABLE IF EXISTS {$wpdb->prefix}b4l_issuer_information" );
     $wpdb->query( "DROP TABLE IF EXISTS {$wpdb->prefix}b4l_userStudentBadgesProfil" );
     $wpdb->query( "DROP TABLE IF EXISTS {$wpdb->prefix}b4l_userTeacherBadgesProfil" );
-//}
+}
 
 //Drop all the custom roles and custom capabilities
 require plugin_dir_path( __FILE__ ) . 'includes/initialisation/users_roles_and_capabilities.php';
